+++ pyproj/_transformer.pyx
@@ -274,7 +274,6 @@ cdef PJ* proj_create_crs_to_crs(
     allow_ballpark,
     bint force_over,
     only_best,
-    allow_superseded,
 ) except NULL:
     """
     This is the same as proj_create_crs_to_crs in proj.h
@@ -534,7 +533,6 @@ cdef class _Transformer(Base):
         allow_ballpark=None,
         bint force_over=False,
         only_best=None,
-        allow_superseded=False,
     ):
         """
         Create a transformer from CRS objects
@@ -577,7 +575,6 @@ cdef class _Transformer(Base):
                 allow_ballpark=allow_ballpark,
                 force_over=force_over,
                 only_best=only_best,
-                allow_superseded=allow_superseded,
             )
         finally:
             if pj_area_of_interest != NULL: