+++ pyproj/transformer.py
@@ -94,7 +94,6 @@ class TransformerFromCRS(  # pylint: disable=too-many-instance-attributes
     allow_ballpark: Optional[bool]
     force_over: bool = False
     only_best: Optional[bool] = None
-    allow_superseded: bool = False
 
     def __call__(self) -> _Transformer:
         """
@@ -112,7 +111,6 @@ class TransformerFromCRS(  # pylint: disable=too-many-instance-attributes
             allow_ballpark=self.allow_ballpark,
             force_over=self.force_over,
             only_best=self.only_best,
-            allow_superseded=self.allow_superseded,
         )
 
 
@@ -561,7 +559,6 @@ class Transformer:
         allow_ballpark: Optional[bool] = None,
         force_over: bool = False,
         only_best: Optional[bool] = None,
-        allow_superseded: bool = False,
     ) -> "Transformer":
         """Make a Transformer from a :obj:`pyproj.crs.CRS` or input used to create one.
 
@@ -575,7 +572,6 @@ class Transformer:
         .. versionadded:: 3.1.0 authority, accuracy, allow_ballpark
         .. versionadded:: 3.4.0 force_over
         .. versionadded:: 3.5.0 only_best
-        .. versionadded:: 3.6.0 allow_superseded
 
         Parameters
         ----------
@@ -619,10 +615,6 @@ class Transformer:
             ``only_best_default`` setting of :ref:`proj-ini`.
             The only_best kwarg overrides the default value if set.
             Requires PROJ 9.2+.
-        allow_superseded: bool, default=False
-            Set to True to allow the use of superseded (but not deprecated)
-            transformations in the candidate coordinate operations. Default is
-            to disallow.
 
         Returns
         -------
@@ -640,7 +632,6 @@ class Transformer:
                 allow_ballpark=allow_ballpark,
                 force_over=force_over,
                 only_best=only_best,
-                allow_superseded=allow_superseded,
             )
         )